# Correspondence_backend/src/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Correspondence System API Documentation",
      "description": "Fake Online REST API for Testing and Prototyping",
      "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/draft/paging/v2": {
            "get": {
                "summary": "Get all drafted threads and drafted correspondences",
                "description": "get all drafts",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    { 
                        "in": "query",
                        "name": "limit",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got all drafted threads successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/draft/draftDocument": {
            "post": {
                "summary": "Draft a particular correspondence",
                "description": "Given a correspondence object, we have to update it in the DB if it exists and mark it as drafted, and if it doesn't exist, we should createa  new one and return it",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "corresondence",
                        "description": "the object containing correspondence data to be drafted",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/correspondence"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Correspondence is drafted successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "doc": {
                                    "type": "object"
                                },
                                "status": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/draft/draftReply": {
            "post": {
                "summary": "Drafting a reply on a particular thread",
                "description": "Given the thread ID and the ID of the parent message, add this object if doesn't exist and mark it as drafted, or if its exist, just mark as drafted",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "parent_id",
                        "description": "The ID of the parent message to attached with",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "thread_id",
                        "description": "The thread id that is common across all of the correspondences",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "doc",
                        "description": "the object containing correspondence data to be added as a reply",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/correspondence"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Correspondence is drafted successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "doc": {
                                    "type": "object"
                                },
                                "status": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/draft/count/v2": {
            "get": {
                "summary": "Get the count of all drafted threads",
                "description": "Get the count of all drafted threads",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    }, 
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "correspondence": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "corr_no": {
                    "type": "string"
                },
                "corr_type": {
                    "type": "string"
                },
                "entity_no": {
                    "type": "string"
                },
                "from_user": {
                    "type": "array"
                },
                "cc_entity": {
                    "type": "array"
                },
                "to_entity": {
                    "type": "string"
                },
                "to_department": {
                    "type": "string"
                },
                "sent_date": {
                    "type": "string"
                },
                "corr_subject": {
                    "type": "string"
                },
                "corr_body": {
                    "type": "string"
                },
                "message_status": {
                    "type": "string"
                },
                "docs_IDs": {
                    "type": "array"
                },
                "due_date": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "draft": {
                    "type": "boolean"
                },
                "thread_id": {
                    "type": "string"
                },
                "await_reply": {
                    "type": "boolean"
                },
                "classification": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "starred": {
                    "type": "boolean"
                },
                "isThreadStarred": {
                    "type": "boolean"
                },
                "isThreadDrafted": {
                    "type": "boolean"
                }
            }
        }
    }
  }